rpid/pyproject.toml
[tool.poetry]
name = "rpid"
description = "Raspbery PI Daemon"
version = "0.0.1"
authors = [ "RGBCube" ]

[tool.poetry.dependencies]
python = "^3.8"

fastapi    = "0.105.0"
"RPi.GPIO" = "0.7.1"
uvicorn    = "0.24.0.post1"
